# back-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aplicativo MiauChei",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Local"
        }
    ],
    "paths": {
...
        "/profile": {
            "get": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Perfil do Usuário",
                "description": "Retorno do perfil do Usuário",
                "operationId": "5d0deb22f4dd2c57642ce790c940c21b",
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Atualização do Perfil",
                "description": "Retorna dados atualizados",
                "operationId": "60b23a2c311e31aca76330467a2b88e5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "byte"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Armando",
                                    "phone": "999999999",
                                    "password": "12345678",
                                    "photo": "U3dhZ2dlciByb2Nrcw=="
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}